requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
flask>=3.0.0
requests>=2.31.0
